lib/checkout: Lock repository during checkout operations
When checking out objects, take a shared lock to ensure objects aren't
deleted. If the repo isn't writable, the locking is a noop. This ensures
that the repository owner can make robust checkouts, but it still allows
other users to make a likely to be successful checkout if the repo mode
supports it. During garbage collection, take an exclusive lock while
deleting uncompressed objects.