lib/repo: Take lock when deleting objects
Ideally this would always be an exclusive lock, but this is also called
when a commit object is written. Taking an exclusive lock there would
prevent commit and pull concurrency. Refactor ostree_repo_delete_object
to allow an internal version that only takes a shared lock and use that
when writing the commit.