bin/admin: Add --lock-timeout option
Admin builtins that require repository locking set the
OSTREE_BUILTIN_FLAG_LOCKING flag to add a --lock-timeout option. This
needs to be used in conjunction with OSTREE_BUILTIN_FLAG_NO_REPO since
the the timeout can only be applied once the repo is gathered from the
sysroot. This allows the user to control how long to wait to acquire the
repository lock.

The cmdprivate version of ostree_repo_set_lock_timeout() is used until
it's no longer experimental.