find-remotes: Add --disable-avahi option
It can be helpful to be able to disable the OstreeRepoFinderAvahi when
using the find-remotes command. For example, if the tests need to run in
an environment that can't have an Avahi daemon, as is the case at
Endless. This commit adds the --disable-avahi option, which will be used
in tests in a following commit.