Add ostree_repo_checkout: Add OSTREE_REPO_CHECKOUT_OVERWRITE_UPDATE
The goal is to do ostree deploys faster by updating already checked out
trees.

This adds an overwrite mode and a new option to specify the original tree.
It will do a lot less work when you're updating a tree with a large number
of files but only a few files have changed.  So if you have a full distro
tree checked out and want to update `/usr/bin/ostree` we no longer have
to recurse through every file in `/usr/share/doc` to do the update.