WIP: rofiles: Implement more FUSE ops
I was debugging a problem with glibc's `build-locale-archive.c` that
boils down to

```
lseek(3, 0, SEEK_END)                   = 103860
```

being wrong.  Digging into this a bit more, I noticed we're not
implementing some of the FUSE ops.

However, now `build-locale-archive` SEGVs!

Looking at this a bit more, I think this is due to the way it does
writes via `mmap()`; FUSE does some caching of the file data, and
that gets out of sync.

There seems to be some support in the latest FUSE 3.0+ to fix this.