tests/str: Migrate some itests to nondestructive/ test group
My main goal is to migrate the some shell script into Ansible (although that
only sort of started), but also split them into "destructive" and
"nondestructive". The nondestructive ones we can in the future run in parallel
on the same VM.

Since the filenames will be new (hence breaking `git annotate`), I made some
effort to track down the original git commit for the shell script test so it's
easy to know why the test was written later.

This also involved some tweaks so that the `nondestructive.yml` and
`sysinstall-tests.yml` playbooks can share more logic.